requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
streamlit>=1.31
SQLAlchemy>=2.0
psycopg2-binary>=2.9
pydantic>=2.6
python-dotenv>=1.0
requests>=2.31
pytest>=8.0
httpx>=0.27
